move utils into separate files
