+++ lib/helpers/index.js
@@ -1,1 +1,6 @@
-module.exports = require('export-files')(__dirname);
+'use strict';
+
+module.exports = {
+  plural: require('./plural'),
+  single: require('./single')
+};
