+++ lib/helpers/plural.js
@@ -1,7 +1,6 @@
 'use strict';
 
 var List = require('../list');
-var utils = require('../utils');
 
 module.exports = function (app, collection, options) {
