+++ lib/item.js
@@ -4,7 +4,7 @@ var Base = require('base-methods');
 var cloneStats = require('clone-stats');
 var Stream = require('stream');
 var Vinyl = require('vinyl');
-var utils = require('./utils');
+var utils = require('./utils/');
 
 /**
  * Create an instance of `Item`. Optionally pass a default object
