+++ lib/utils/lazy.js
@@ -6,7 +6,10 @@
 
 var lazy = require('lazy-cache')(require);
 
-// object/array/type utils
+/**
+ * Common utils
+ */
+
 lazy('clone');
 lazy('is-buffer');
 lazy('paginationator');
@@ -16,10 +19,16 @@ lazy('define-property', 'define');
 lazy('mixin-deep', 'merge');
 lazy('extend-shallow', 'extend');
 
-// routing
+/**
+ * Routing
+ */
+
 lazy('en-route', 'router');
 
-// engines, templates and helpers
+/**
+ * Engines, templates and helpers utils
+ */
+
 lazy('load-helpers', 'loader');
 lazy('engine-base', 'engine');
 lazy('engine-cache', 'Engines');
@@ -28,10 +37,17 @@ lazy('template-error', 'rethrow');
 lazy('inflection', 'inflect');
 lazy('layouts');
 
-// glob/matching utils
+/**
+ * Globbing and matching utils
+ */
+
 lazy('globby');
 lazy('micromatch', 'mm');
 lazy('is-valid-glob');
 lazy('has-glob');
+
+/**
+ * Expose utils
+ */
 
 module.exports = lazy;
