+++ lib/view.js
@@ -1,9 +1,5 @@
 'use strict';
 
-var Base = require('base-methods');
-var cloneStats = require('clone-stats');
-var Stream = require('stream');
-var Vinyl = require('vinyl');
 var utils = require('./utils');
 var Item = require('./item');
